Refactor Docker SSH implementation and update UI components
- Replace old SSH client with new dockerssh package
- Update volume service and related domain objects
- Refactor UI builders and handlers for better separation
- Update logger and error handling throughout the application
- TODO: lint and tests
