+++ internal/config/config_test.go
@@ -80,7 +80,7 @@ func TestLoad_ExistingConfig(t *testing.T) {
 	}()
 
 	configDir := filepath.Join(tempHome, ".dockerk9s")
-	err := os.MkdirAll(configDir, 0o755)
+	err := os.MkdirAll(configDir, 0o750)
 	require.NoError(t, err)
 
 	configFile := filepath.Join(configDir, "config.json")
@@ -91,7 +91,7 @@ func TestLoad_ExistingConfig(t *testing.T) {
 		"docker_host": "tcp://localhost:2375",
 		"theme": "dark"
 	}`
-	err = os.WriteFile(configFile, []byte(testConfig), 0o644)
+	err = os.WriteFile(configFile, []byte(testConfig), 0o600)
 	require.NoError(t, err)
 
 	cfg, err := Load()
@@ -125,7 +125,7 @@ func TestLoad_InvalidConfig(t *testing.T) {
 	}()
 
 	configDir := filepath.Join(tempHome, ".dockerk9s")
-	err := os.MkdirAll(configDir, 0o755)
+	err := os.MkdirAll(configDir, 0o750)
 	require.NoError(t, err)
 
 	configFile := filepath.Join(configDir, "config.json")
@@ -134,7 +134,7 @@ func TestLoad_InvalidConfig(t *testing.T) {
 		"refresh_interval": "invalid",
 		"log_level": "DEBUG"
 	`
-	err = os.WriteFile(configFile, []byte(invalidConfig), 0o644)
+	err = os.WriteFile(configFile, []byte(invalidConfig), 0o600)
 	require.NoError(t, err)
 
 	cfg, err := Load()
@@ -216,7 +216,7 @@ func TestThemeLoading(t *testing.T) {
   error: "red"
   info: "cyan"`
 
-	err := os.WriteFile(themePath, []byte(themeContent), 0o644)
+	err := os.WriteFile(themePath, []byte(themeContent), 0o600)
 	require.NoError(t, err)
 
 	tm := NewThemeManager(themePath)
