+++ internal/ui/interfaces/ui_interface.go
@@ -1,3 +1,4 @@
+//nolint:revive // Package name "interfaces" is for interface-only package
 package interfaces
 
 import (
