+++ main.go
@@ -1,3 +1,4 @@
+// Package main is the entry point for the WhaleTUI application.
 package main
 
 import "github.com/wikczerski/whaletui/cmd"
